refactor: break common to non-common dep in analyzer-telemetry-callbacks (#3715)
#### Description of changes

This PR fixes one of the big reasons `yarn null:find-cycles` finds so many results; specifically, it breaks a dependency from `common/types/analyzer-telemetry-callbacks.ts` to `injected/analyzers/analyzer.ts` by moving the `AxeAnalyzerResult` type out of `analyzer.ts` into its own file under `common/types`.

Previously, this dependency was creating *tons* of cycles (essentially, any code that touched anything telemetry related transitively included the telemetry callbacks, which then ended up going through analyzers implementation, which involves some telemetry).

Line counts of `yarn null:find-cycles` (reduced by 38% with this PR):

Before:
```
C:\repos\accessibility-insights-web [master ≡] | 59 | 14:34:24 12/04
> yarn null:find-cycles | measure
Count             : 15839
```

After:
```
C:\repos\accessibility-insights-web [break-axe-analyzer-result-cycle ≡] | 54 | 14:33:17 12/04
> yarn null:find-cycles | measure
Count             : 9850
```

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->
- [x] Addresses an existing issue: unblocks #2869
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [n/a] Verified code coverage for the changes made. Check coverage report at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See `CONTRIBUTING.md`.
- [n/a] (UI changes only) Added screenshots/GIFs to description above
- [n/a] (UI changes only) Verified usability with NVDA/JAWS
