+++ src/injected/analyzers/tab-stops-analyzer.ts
@@ -1,17 +1,13 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 import { Logger } from 'common/logging/logger';
+import { AxeAnalyzerResult } from 'common/types/axe-analyzer-result';
 import { WindowUtils } from 'common/window-utils';
 import { ScanIncompleteWarningDetector } from 'injected/scan-incomplete-warning-detector';
 import * as Q from 'q';
 
 import { TabStopEvent, TabStopsListener } from '../tab-stops-listener';
-import {
-    AxeAnalyzerResult,
-    FocusAnalyzerConfiguration,
-    ScanBasePayload,
-    ScanUpdatePayload,
-} from './analyzer';
+import { FocusAnalyzerConfiguration, ScanBasePayload, ScanUpdatePayload } from './analyzer';
 import { BaseAnalyzer } from './base-analyzer';
 
 export interface ProgressResult<T> {
