+++ src/injected/analyzers/unified-result-sender.ts
@@ -1,6 +1,9 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
+import { UnifiedScanCompletedPayload } from 'background/actions/action-payloads';
 import { ScanIncompleteWarningsTelemetryData } from 'common/extension-telemetry-events';
+import { Messages } from 'common/messages';
+import { AxeAnalyzerResult } from 'common/types/axe-analyzer-result';
 import { ToolData } from 'common/types/store-data/unified-data-interface';
 import { FilterResults } from 'injected/analyzers/filter-results';
 import {
@@ -10,14 +13,12 @@ import {
 import { ScanIncompleteWarningDetector } from 'injected/scan-incomplete-warning-detector';
 import { isEmpty } from 'lodash';
 import { ScanResults } from 'scanner/iruleresults';
-import { UnifiedScanCompletedPayload } from '../../background/actions/action-payloads';
-import { Messages } from '../../common/messages';
+
 import {
     ConvertScanResultsToUnifiedResults,
     ConvertScanResultsToUnifiedResultsDelegate,
 } from '../adapters/scan-results-to-unified-results';
 import { ConvertScanResultsToUnifiedRulesDelegate } from '../adapters/scan-results-to-unified-rules';
-import { AxeAnalyzerResult } from './analyzer';
 import { MessageDelegate, PostResolveCallback } from './rule-analyzer';
 
 export class UnifiedResultSender {
